+++ src/app/main_window.rs
@@ -1,12 +1,12 @@
-use crate::lineeditor::{ EditorEvent, convert_keyevent_to_editorevent };
+use crate::app::key_select_menu::KeySelectMenu;
+use crate::app::{App, CachedCommandPart, KeySelectMenuType};
+use crate::lineeditor::{convert_keyevent_to_editorevent, EditorEvent};
+use crate::util::{StringExt, VecStringExt};
 use crate::CmdOutput;
 use crate::Stdio;
-use crate::app::key_select_menu::KeySelectMenu;
-use crate::util::{StringExt, VecStringExt};
-use crate::app::app::{App, KeySelectMenuType, CachedCommandPart};
 
-use std::path::Path;
 use crossterm::event::{KeyCode, KeyModifiers};
+use std::path::Path;
 use std::path::PathBuf;
 use std::process::Command;
